addes tests for: FakeFrameGrabber, frameGrabberCropper, frameGrabber_nws_yarp, frameGrabber_nwc_yarp, grabberDual
