+++ src/devices/frameGrabberCropper/tests/frameGrabberCropperTest.cpp
