+++ src/devices/frameGrabber_nws_yarp/CMakeLists.txt
@@ -46,4 +46,9 @@ if(NOT SKIP_frameGrabber_nws_yarp)
   set(YARP_${YARP_PLUGIN_MASTER}_PRIVATE_DEPS ${YARP_${YARP_PLUGIN_MASTER}_PRIVATE_DEPS} PARENT_SCOPE)
 
   set_property(TARGET yarp_frameGrabber_nws_yarp PROPERTY FOLDER "Plugins/Device/NWS")
+
+  if(YARP_COMPILE_TESTS)
+    add_subdirectory(tests)
+  endif()
+
 endif()
