+++ src/libYARP_dev/src/yarp/dev/tests/IRgbVisualParamsTest.cpp
@@ -1,0 +1,6 @@
+/*
+ * SPDX-FileCopyrightText: 2006-2021 Istituto Italiano di Tecnologia (IIT)
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+#include "IRgbVisualParamsTest.h"
